requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.3.0

# Experiment Tracking
mlflow>=2.8.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
opencv-python>=4.8.0

# Utilities
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
prometheus-client>=0.19.0

torch
torchvision
